requirements-alphastar.txt
ceia-soccer-twos
dm-tree==0.1.6
gym==0.19.0
gym-unity==0.27.0
mlagents==0.27.0
mlagents-envs==0.27.0
numpy==1.19.5
ray==1.8.0
ray[tune]==1.8.0
ray[rllib]==1.8.0